// run-archive/scenario-1/1778850600043-scenario-1-deepseek-v4-pro-plan-flash-build-mimo-review-attempt-1/frontend/src/api/generated/schemas/restaurantResponse.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * RestaurantBooking.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface RestaurantResponse {
  id: string;
  name: string;
  description: string;
  cuisine: string;
  address: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  tableCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  minCapacity: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  maxCapacity: number | string;
}